Migrate from two databases to one
Database tables are moved to the modeemiuserdb
and models with constraints have been added to code base,
which enables easier testing and creates similar tables
to the local environment as in production environment.

This allows for simpler source code and database semantics.

The models have been generated with inspectdb from modeemiuserdb
and models from modeemintternet database have been migrated to the
modeemiuserdb which will be used by all applications from now on.

The models have been diffed manually and are similar to the
actual production modeemiuserdb models aside from a few
automatically managed search indexes which do not affect behaviour.